pyproject.toml
[project]
name = "hirestream-match"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "dashscope>=1.25.0",
    "fastapi>=0.115.0",
    "loguru>=0.7.3",
    "markupsafe==3.0.3",
    "multidict==6.7.0",
    "narwhals==2.10.2",
    "networkx==3.4.2",
    "nibabel==5.3.2",
    "nipype==1.10.0",
    "numpy==2.2.6",
    "openai==2.7.1",
    "packaging==25.0",
    "pandas==2.3.3",
    "pathlib==1.0.1",
    "pillow==12.0.0",
    "propcache==0.4.1",
    "protobuf==6.33.0",
    "prov==2.1.1",
    "puremagic==1.30",
    "pyarrow==21.0.0",
    "pycparser==2.23",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydeck==0.9.1",
    "pydot==4.0.1",
    "pyjwt>=2.10.1",
    "pymupdf==1.26.6",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyxnat==1.6.4",
    "rdflib==7.4.0",
    "referencing==0.37.0",
    "requests==2.32.5",
    "rpds-py==0.28.0",
    "scipy==1.15.3",
    "simplejson==3.20.2",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.44",
    "starlette==0.50.0",
    "streamlit==1.51.0",
    "gradio>=4.44.0",
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traits==7.0.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "watchdog==6.0.0",
    "websocket-client==1.9.0",
    "win32-setctime==1.2.0",
    "yarl==1.22.0",
]
